requirements.txt
certifi==2022.5.18.1
charset-normalizer==2.0.12
greenlet==1.1.2
idna==3.3
numpy==1.22.4
pandas==1.4.2
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
requests==2.27.1
six==1.16.0
spotipy==2.19.0
SQLAlchemy==1.4.37
urllib3==1.26.9
